requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
cent==5.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-bootstrap-v5==1.0.11
django-bootstrap5==24.1
Faker==24.7.1
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
multidict==6.0.5
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
requests==2.32.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
types-requests==2.32.0.20240523
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
